Проведены координальные изменения во всех сущностях. Созданы новые сущности, репозитории, сервисы и контроллеры для счетов, счетчиков, показаний счетчиков, реквизитов компании. Изменены шаблоны страниц. Изменена структура БД

+++ src/main/java/ru/geekbrains/domodel/repositories/MeterRepository.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Repository;
 import ru.geekbrains.domodel.entities.Meter;
 
 /**
- * Репозиторий счетчиков
+ * Репозиторий счетчиков показаний
  */
 @Repository
-public interface MeterRepository extends JpaRepository<Meter, Integer> {
+public interface MeterRepository extends JpaRepository<Meter, Long> {
 
 }
